requirements.txt
requests
torchsummary
torchvision
APScheduler
Flask_RESTful
Flask
torch
Flask_Cors
numpy
Pillow
beautifulsoup4
websockets
opencv-contrib-python
scikit-image
imutils
selenium==3.141.0
selenium-wire==1.0.11
furl
retry==0.9.2